packages/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
httpx==0.28.1
python-multipart==0.0.20
stripe==11.4.1
boto3==1.36.4
resend==2.7.0
tiktoken==0.8.0
slowapi==0.1.9
pyyaml==6.0.2
tapo==0.8.12
